Extrinsics calibration node is now functional
Added the detect_tags function to the apriltag pose estimation library. It ensure the corners are in the correct order for IPPE PnP solving.

Added the get_marker_pose function to the apriltag pose estimation library. It solves for the position of the marker in the VICON frame given the pose of a tag and enough information to solve for the relative position of the marker with respect to the tag.

Created the /data/final_calibration folder to hold calibration files that have been tested and shown to be correct

Changed tag_size everywhere to be in meters so we don't have to keep switching back and forth

Updated calibrate.py to reproject corners onto the image to vizualize the RMSE