Create a second version of the vicon_set_position node
This one keeps track of a transform from the VICON frame to the local/realsense frame so that we can transform positions from realsense back to vicon for the purpose of checking whether we have gotten to the location and visualization. Is also has the ability to use localization from the apriltag localizer instead of ground truth VICON. In order to facilitate this, I also added a TagTransformArray message type that is used by the apriltag localization node to send information about our predicted position.

I also added a node that will record flight data for the purpose of visualization later. It keeps track of our ground truth position, our corrected position using vicon_set_position, and the uncorrected realsense position.