Added path planning
The planner uses networkx to find the minimum distance route to all tags in a AprilTagMap object

Added a set_position service to the comms node for manually flying the drone. Updated the ground station to be able to use this new service.

Added mapping logic to comms node. It also tells the vicon set position node to switch to VICON localization at the start.

Added inspecting logic to the comms node. This one tells the vicon set position node to switch to AprilTag localization at the start.